removed the tacky hover stuff and just make this a formatting improvement for now
